Use more efficient method to generate genesis state
