+++ eth2/state_processing/Cargo.toml
@@ -32,6 +32,7 @@ tree_hash = "0.1.0"
 tree_hash_derive = "0.2"
 types = { path = "../types" }
 rayon = "1.2.0"
+eth2_hashing = { path = "../utils/eth2_hashing" }
 
 [features]
 fake_crypto = ["bls/fake_crypto"]
